feat: change root layout, add theme provider
