+++ src/assets/style/mixins.scss
@@ -18,7 +18,7 @@
   @include center-flex;
   width: $width;
   text-decoration: none;
-  font-size: .8rem;
+  font-size: 0.8rem;
   height: $height;
   border: none;
   border-radius: 4px;
