+++ src/components/character-detailed/character-detailed.module.scss
@@ -59,9 +59,9 @@
     border: 2px solid $c-main-bg;
     background-color: red;
     color: white;
-      &:hover {
-        background-color: darkred !important;
-      }
+    &:hover {
+      background-color: darkred !important;
+    }
   }
 }
