+++ src/components/character-detailed/detailed-error/detailed-error.module.scss
