+++ src/components/search/search.module.scss
@@ -38,7 +38,6 @@
   padding: 20px;
 }
 
-
 [data-theme='light'] {
   .search-results {
     background-color: $c-light-page-bg;
