+++ src/components/ui/button/button.module.scss
@@ -3,7 +3,6 @@
 
 .button {
   @include button(120px, 40px);
-
 }
 
 .error-button {
